сохраняемся.. количество подключений к фтп ограничено, решено sqlite3.OperationalError: database is locked, надо решить

+++ app.py
@@ -13,7 +13,7 @@ host, port, login, password = 'ftp.zakupki.gov.ru', 21, 'free', 'free'
 links = []
 folders = [
     '/fcs_regions/Tulskaja_obl/contracts/currMonth',
-    # '/fcs_regions/Tulskaja_obl/contracts/prevMonth'
+    '/fcs_regions/Tulskaja_obl/contracts/prevMonth'
 ]
 
 
@@ -61,66 +61,75 @@ async def not_exist_in_db(ftp_path: str, modify: str):
     return not bool(res[0])
 
 
-async def get_data(ftp_path: str, modify: str):
+async def get_data(ftp_path: str, modify: str, sem):
     file = ftp_path.split('/')[-1]
-    while True:
-        try:
-            async with aioftp.Client.context(host, port, login, password) as client:
-                print(f"Downloading file {file}...")
-                await client.download(ftp_path, f"Temp/{file}", write_into=True)
-                print(f"Finished downloading file {file} into Temp/{file}")
-            break
-        except ConnectionResetError:
-            print('ConnectionResetError')
-            pass
+    async with sem:
+        while True:
+            try:
+                async with aioftp.Client.context(host, port, login, password) as client:
+                    print(f"Downloading file {file}...")
+                    await client.download(ftp_path, f"Temp/{file}", write_into=True)
+                    print(f"Finished downloading file {file} into Temp/{file}")
+                break
+            except ConnectionResetError:
+                print('ConnectionResetError')
+                pass
     event_data = []
-    if file == 'contract_Tulskaja_obl_2024011200_2024011300_001.xml.zip':
-        z = zipfile.ZipFile(f'Temp//{file}', 'r')
+    # if file == 'contract_Tulskaja_obl_2024011200_2024011300_001.xml.zip':
+    with zipfile.ZipFile(f'Temp//{file}', 'r') as z:
         for item in z.namelist():
-            if item.endswith('.xml'):
+            if item.endswith('.xml') and not any([item.startswith('contractAvailableForElAct'), item.startswith('contractProcedureCancel')]):
                 print(f'Extract {item} from {ftp_path}')
                 z.extract(item, 'Temp')
                 with open(f'Temp//{item}') as f:
                     src = f.read()
-                    print(re.search(r'(?<=<regNum>)\d{19}(?=</regNum>)', src)[0],
-                          re.search(r'(?<=<publishDate>).+(?=</publishDate>)', src)[0])
+                    if item.startswith('contract'):
+                        eisdocno = re.search(r'(?<=<regNum>)\d{19}(?=</regNum>)', src)[0]
+                    try:
+                        eispublicationdate = re.search(r'(?<=<publishDate>).+(?=</publishDate>)', src)[0]
+                    except Exception as e:
+                        print(e, item)
+                    print(eisdocno, eispublicationdate)
                     event_data.append({
                         'ftp_path': ftp_path,
                         'modify': modify,
-                        'eisdocno': re.search(r'(?<=<regNum>)\d{19}(?=</regNum>)', src)[0],
-                        'eispublicationdate': re.search(r'(?<=<publishDate>).+(?=</publishDate>)', src)[0],
+                        'eisdocno': eisdocno,
+                        'eispublicationdate': eispublicationdate,
                         'xmlname': item
                     })
                 os.unlink(f'Temp//{item}')
-        z.close()
     for row in event_data:
         await insert_event(row)
 
 
-async def get_ftp_list(folder: str):
-    async with aioftp.Client.context(host, port, login, password) as client:
-        x = await client.list(folder, recursive=False)
-    for path, info in x:
-        print(path, info)
-        no_rows_in_db = await not_exist_in_db(str(path), info['modify'])
-        print('Записей в базе нет:', no_rows_in_db)
-        if info['size'] != '22' and no_rows_in_db:
-            links.append(
-                (str(path), info['modify'])
-            )
+async def get_ftp_list(folder: str, sem):
+    async with sem:
+        async with aioftp.Client.context(host, port, login, password) as client:
+            x = await client.list(folder, recursive=False)
+        for path, info in x:
+            print(path, info)
+            no_rows_in_db = await not_exist_in_db(str(path), info['modify'])
+            print('Записей в базе нет:', no_rows_in_db)
+            if info['size'] != '22' and no_rows_in_db:
+                links.append(
+                    (str(path), info['modify'])
+                )
 
 
 async def main():
+
     if not os.path.exists('Temp'):
         os.mkdir('Temp')
+
     await create_tables()
 
+    semaphore = asyncio.Semaphore(50)
     tasks = [
-        asyncio.create_task(get_ftp_list(folder)) for folder in folders
+        asyncio.create_task(get_ftp_list(folder, semaphore)) for folder in folders
     ]
     await asyncio.gather(*tasks)
     tasks = [
-        asyncio.create_task(get_data(ftp_path, modify)) for ftp_path, modify in links
+        asyncio.create_task(get_data(ftp_path, modify, semaphore)) for ftp_path, modify in links
     ]
     await asyncio.gather(*tasks)
